Add initComponent
